Test copying from server, running on local13

+++ Rstox_utils.R
@@ -960,7 +960,9 @@ automatedRstoxTest <- function(dir, root=list(windows="\\\\delphi", unix="/Volum
 		out <- lapply(files$commonFiles, diffTextFilesOne, dir1=files$dir1, dir2=files$dir2, progressFile=progressFile, diffdir=diffdir, nlines=nlines)
 		write("}", file=progressFile, append=TRUE)
 		
-		unlink(tempdiff)
+		if(file.exists(tempdiff)){
+			unlink(tempdiff)
+		}
 		
 		return(NULL)
 	}
